Organize Task, ToDo, Deadline, Event and TaskList classes into nyx.task package
