+++ src/main/java/nyx/task/Event.java
@@ -1,4 +1,4 @@
-package nyx;
+package nyx.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
