+++ src/main/java/nyx/task/Task.java
@@ -1,4 +1,4 @@
-package nyx;
+package nyx.task;
 
 /**
  * Represents a task.
