+++ src/test/java/nyx/DeadlineTest.java
@@ -2,6 +2,7 @@ package nyx;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import nyx.task.Deadline;
 import org.junit.jupiter.api.Test;
 
 public class DeadlineTest {
